+++ R/poisson.R
@@ -174,7 +174,7 @@ poisson_test_pv <- function(
 
     res[idx] <- pv_supp[x[idx] + 1]
     if(!simple_output) {
-      supports[[i]] <- sort(unique(pv_supp))
+      supports[[i]] <- unique(sort(pv_supp))
       indices[[i]]  <- idx
     }
   }