+++ NEWS.md
@@ -1,4 +1,13 @@
-## DiscreteTests 0.2.0
+# DiscreteTests 0.2.1
+
+* `get_pvalues()` method of `DiscreteTestResults` class now includes the (row or
+  cell) names of the input observations (if present). Output of names can be
+  suppressed by a parameter.
+* `print()` method of `DiscreteTestResults` class now also prints the (row or
+  cell) names of the input observations (if present).
+
+
+# DiscreteTests 0.2.0
 
 * Migration to snake case for all function and parameter names. Previously
   existing functions are kept for compatibility, but are marked as deprecated
@@ -6,20 +15,21 @@
 * `print()` output of `DiscreteTestResults` objects now wraps output of all 
   input data, parameters and hypotheses correctly
 
-## DiscreteTests 0.1.3
+
+# DiscreteTests 0.1.3
 
 * Minor bug fix for `fisher.test.pv()` and  `mcnemar.test.pv()` functions:
   malformed input data was sometimes still accepted, but caused errors later
 
 
-## DiscreteTests 0.1.2
+# DiscreteTests 0.1.2
 
 * Minor improvements of `print()` method in `DiscreteTestResults` class
 * Additions to examples in `DiscreteTestResultsSummary` class and
   `summary.DiscreteTestResults()` function
 
 
-## DiscreteTests 0.1.1
+# DiscreteTests 0.1.1
 
 * Improved `print()` method in `DiscreteTestResults` class
   - long lines are now wrapped and indented with `strwrap()`
@@ -30,7 +40,7 @@
 * Updated references
 
 
-## DiscreteTests 0.1.0
+# DiscreteTests 0.1.0
 
 * Initial release.
 * Added a `NEWS.md` file to track changes to the package.